Feat : 로또 번호 관련 기능 lotteryController에 추가
